backend/requirements.txt
transformers
flask
beautifulsoup4
requests
SentencePiece
torch
pymongo 
Flask-PyMongo 
websocket
tradingview-ta
accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
dnspython==2.6.1
exceptiongroup==1.2.2
executing==2.0.1
faiss-cpu==1.8.0.post1
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-PyMongo==2.3.0
frozenlist==1.4.1
fsspec==2024.6.1
gevent==24.2.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.1
grpcio-status==1.62.2
httplib2==0.22.0
huggingface==0.0.1
huggingface-hub==0.23.4
idna==3.7
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-google-genai==1.0.7
langchain-text-splitters==0.2.2
langsmith==0.1.93
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
# price-loaders @ git+https://github.com/batprem/price-loaders.git@e796292e858851e29aa74d084505db9607f784f8
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.4
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
sympy==1.13.0
tbb==2021.13.0
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
tradingview-ta==3.3.0
traitlets==5.14.3
transformers==4.42.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
websocket==0.2.1
Werkzeug==3.0.3
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2